adding custom widgets and templates for those widgets, and js for more dinamyc use of the views, changing the clothe model and adding a new models, one called images_and_color_for_clothe and the other image_for_clothe

+++ Simple_Store_D/car_shop/views.py
@@ -1,4 +1,5 @@
 from django.http import HttpRequest,JsonResponse
+from django.shortcuts import redirect
 from .car_shop import Car_Shop
 from . import models
 from .repository import car_shop as car_shop_repo
@@ -25,33 +26,47 @@ def get_clothe_id_color_id_size_id(string_id):
 
 def add_to_car(request:HttpRequest,string_id):
     
-    ids = get_clothe_id_color_id_size_id(string_id)
+    if request.user.is_authenticated:
+        ids = get_clothe_id_color_id_size_id(string_id)
+        
+        car_shop = Car_Shop(request)
+        
+        
+        
+        clothe = general.get_clothe(ids[0])
+        color = general.get_color(color_id=ids[1])
+        size = general.get_size(size_id=ids[2])
+        
+        images = clothe.ColorImages.filter(color=color).first().images.all()
+        
+        image_url = ''
+        for image in images.all():
+            if image_url == '':
+                image_url=image.image.url
+                
+            else:
+                break
+        
+        
+        
+        car_shop.agregate_to_car(clothe,color,size,image_url)
+        
+        order_key = string_id 
+        
+        return JsonResponse(data = {"response":"V",
+                                    "clothe_id":ids[0],
+                                    "units":car_shop.car_shop[order_key]['units'],
+                                    "color_id":car_shop.car_shop[order_key]['color_id'],
+                                    "color" : car_shop.car_shop[order_key]['color'],
+                                    "size_id" : car_shop.car_shop[order_key]['size_id'],
+                                    "size": car_shop.car_shop[order_key]['size'],
+                                    "desc":car_shop.car_shop[order_key]['description'],
+                                    "price":car_shop.car_shop[order_key]['price'],
+                                    "image":car_shop.car_shop[order_key]['img'],
+                                    "items_id":list(car_shop.car_shop.keys())})
     
-    car_shop = Car_Shop(request)
-    
-    
-    
-    clothe = general.get_clothe(ids[0])
-    color = general.get_color(color_id=ids[1])
-    size = general.get_size(size_id=ids[2])
-    
-    
-    
-    car_shop.agregate_to_car(clothe,color,size)
-    
-    order_key = string_id 
-    
-    return JsonResponse(data = {"response":"V",
-                                "clothe_id":ids[0],
-                                "units":car_shop.car_shop[order_key]['units'],
-                                "color_id":car_shop.car_shop[order_key]['color_id'],
-                                "color" : car_shop.car_shop[order_key]['color'],
-                                "size_id" : car_shop.car_shop[order_key]['size_id'],
-                                "size": car_shop.car_shop[order_key]['size'],
-                                "desc":car_shop.car_shop[order_key]['description'],
-                                "price":car_shop.car_shop[order_key]['price'],
-                                "items_id":list(car_shop.car_shop.keys())})
-
+    return redirect('login')
+        
 
 def low_in_car(request:HttpRequest,string_id):
     
@@ -93,15 +108,25 @@ def delete_in_car(request:HttpRequest,string_id):
 
 
 def clear_car(request:HttpRequest):
-    car_shop = Car_Shop(request)
     
-    car_shop.delete_all()
+    if request.user.is_authenticated:
+        
+        car_shop = Car_Shop(request)
+        
+        car_shop.delete_all()
+        
+        return JsonResponse(data={"response":"V"})
     
-    return JsonResponse(data={"response":"V"})
+    return redirect('login')
     
 def create_order(request:HttpRequest):
-    car_shop = Car_Shop(request)
-        
-    response = car_shop_repo.create_order(car_shop,request.user)
     
-    return JsonResponse(data=response)
+    if request.user.is_authenticated:
+        
+        car_shop = Car_Shop(request)
+            
+        response = car_shop_repo.create_order(car_shop,request.user)
+        
+        return JsonResponse(data=response)
+    
+    return redirect('login')
